+++ src/engine/Object.cpp
@@ -37,7 +37,7 @@ const std::shared_ptr<ConstantBuffer>& ModelObject::GetBoneMatBuff()
 	{
 		if (!boneBuff)
 		{
-			const int boneNum = model->skelton->bones.size();
+			const int boneNum = static_cast<int>(model->skelton->bones.size());
 			boneBuff = D3D12App::Instance()->GenerateConstantBuffer(sizeof(Matrix), boneNum, nullptr, (model->header.fileName + " - ModelObject - BoneMat").c_str());
 		}
 		boneBuff->Mapping(animator->GetBoneMatricies().data());
